move di to config
