+++ main.go
@@ -9,7 +9,7 @@ import (
 	"github.com/ride-app/driver-service/api/gen/ride/driver/v1alpha1/driverv1alpha1connect"
 	"github.com/ride-app/driver-service/api/interceptors"
 	"github.com/ride-app/driver-service/config"
-	"github.com/ride-app/driver-service/di"
+	"github.com/ride-app/driver-service/config/di"
 	"github.com/ride-app/driver-service/utils/logger"
 	"golang.org/x/net/http2"
 	"golang.org/x/net/http2/h2c"
